+++ packages/core/src/indexing/postgres/__tests__/postgres-index-api.test.ts
@@ -4,7 +4,10 @@ import { StreamID } from '@ceramicnetwork/streamid'
 import knex, { Knex } from 'knex'
 import pgSetup from '@databases/pg-test/jest/globalSetup'
 import pgTeardown from '@databases/pg-test/jest/globalTeardown'
-import { asTableName } from '../../as-table-name.util'
+import { asTableName } from '../../as-table-name.util.js'
+import { IndexQueryNotAvailableError } from '../../index-query-not-available.error.js'
+import { listMidTables } from '../init-tables.js'
+import { Model } from '@ceramicnetwork/stream-model'
 
 const STREAM_ID_A = 'kjzl6cwe1jw145m7jxh4jpa6iw1ps3jcjordpo81e0w04krcpz8knxvg5ygiabd'
 const STREAM_ID_B = 'kjzl6cwe1jw147dvq16zluojmraqvwdmbh61dx9e0c59i344lcrsgqfohexp60s'
@@ -25,25 +28,12 @@ beforeEach(async () => {
 })
 
 afterEach(async () => {
-  await dropMidTables()
+  await dropTables()
   await dbConnection.destroy()
 })
 
-export async function listMidTables() {
-  const dataArr = []
-  await dbConnection
-    .select('tablename')
-    .from('pg_tables')
-    .whereRaw("schemaname='public' AND tablename LIKE ('kjz%')")
-    .then(function (result) {
-      result.forEach(function (value) {
-        dataArr.push(value.tablename)
-      })
-    })
-  return dataArr
-}
-
-export async function dropMidTables() {
+export async function dropTables() {
+  await dbConnection.schema.dropTableIfExists(Model.MODEL.toString())
   await dbConnection.schema.dropTableIfExists(STREAM_ID_A)
   await dbConnection.schema.dropTableIfExists(STREAM_ID_B)
 }
@@ -56,28 +46,62 @@ describe('init', () => {
   describe('create tables', () => {
     test('create new table from scratch', async () => {
       const modelsToIndex = [StreamID.fromString(STREAM_ID_A)]
-      const indexApi = new PostgresIndexApi(dbConnection, modelsToIndex)
+      const indexApi = new PostgresIndexApi(dbConnection, modelsToIndex, true)
       await indexApi.init()
-      const created = await listMidTables()
+      const created = await listMidTables(dbConnection)
       const tableNames = modelsToIndex.map(asTableName)
       expect(created).toEqual(tableNames)
+    })
+
+    test('table creation is idempotent', async () => {
+      const modelsToIndex = [Model.MODEL, StreamID.fromString(STREAM_ID_A)]
+      const indexApi = new PostgresIndexApi(dbConnection, modelsToIndex, true)
+      await indexApi.init()
+      // init again to make sure we don't error trying to re-create the tables
+      await indexApi.init()
+      const created = await listMidTables(dbConnection)
+      const tableNames = modelsToIndex.map(asTableName)
+      expect(created).toEqual(tableNames)
+    })
+
+    test('create new table from scratch but fail table validation', async () => {
+      const INVALID_TABLE_STRUCTURE = {
+        stream_id: {
+          type: 'character varying',
+          maxLength: 254,
+          nullable: false,
+          defaultValue: null,
+        },
+      }
+
+      const modelsToIndex = [StreamID.fromString(STREAM_ID_A)]
+      const indexApi = new PostgresIndexApi(dbConnection, modelsToIndex, true)
+      await indexApi.init()
+      const created = await listMidTables(dbConnection)
+      const tableNames = modelsToIndex.map(asTableName)
+      expect(created).toEqual(tableNames)
+      await expect(indexApi.verify(INVALID_TABLE_STRUCTURE)).rejects.toThrow(
+        /Schema verification failed for index/
+      )
     })
 
     test('create new table with existing ones', async () => {
       // First init with one model
       const modelsA = [StreamID.fromString(STREAM_ID_A)]
-      const indexApiA = new PostgresIndexApi(dbConnection, modelsA)
+      const indexApiA = new PostgresIndexApi(dbConnection, modelsA, true)
       await indexApiA.init()
-      const createdA = await listMidTables()
+      const createdA = await listMidTables(dbConnection)
       const tableNamesA = modelsA.map(asTableName)
       expect(createdA).toEqual(tableNamesA)
 
       // Next add another one
       const modelsB = [...modelsA, StreamID.fromString(STREAM_ID_B)]
-      const indexApiB = new PostgresIndexApi(dbConnection, modelsB)
+      const indexApiB = new PostgresIndexApi(dbConnection, modelsB, true)
       await indexApiB.init()
-      const createdB = await listMidTables()
+      const createdB = await listMidTables(dbConnection)
       const tableNamesB = modelsB.map(asTableName)
+      createdB.sort()
+      tableNamesB.sort()
       expect(createdB).toEqual(tableNamesB)
     })
   })
@@ -88,7 +112,7 @@ describe('close', () => {
     const fauxDbConnection = {
       destroy: jest.fn(),
     } as unknown as Knex
-    const indexApi = new PostgresIndexApi(fauxDbConnection, [])
+    const indexApi = new PostgresIndexApi(fauxDbConnection, [], true)
     await indexApi.close()
     expect(fauxDbConnection.destroy).toBeCalled()
   })
@@ -115,7 +139,7 @@ describe('indexStream', () => {
 
   let indexApi: PostgresIndexApi
   beforeEach(async () => {
-    indexApi = new PostgresIndexApi(dbConnection, MODELS_TO_INDEX)
+    indexApi = new PostgresIndexApi(dbConnection, MODELS_TO_INDEX, true)
     await indexApi.init()
   })
 
@@ -158,5 +182,25 @@ describe('indexStream', () => {
     expect(closeDates(updatedAt, updateTime)).toBeTruthy()
     const createdAt = new Date(raw.created_at)
     expect(closeDates(createdAt, createTime)).toBeTruthy()
+  })
+})
+
+describe('page', () => {
+  const FAUX_DB_CONNECTION = {} as unknown as Knex
+
+  test('call the order if historical sync is allowed', async () => {
+    const indexApi = new PostgresIndexApi(FAUX_DB_CONNECTION, [], true)
+    const mockPage = jest.fn(async () => {
+      return { edges: [], pageInfo: { hasNextPage: false, hasPreviousPage: false } }
+    })
+    indexApi.insertionOrder.page = mockPage
+    await indexApi.page({ model: STREAM_ID_A, first: 100 })
+    expect(mockPage).toBeCalled()
+  })
+  test('throw if historical sync is not allowed', async () => {
+    const indexApi = new PostgresIndexApi(FAUX_DB_CONNECTION, [], false)
+    await expect(indexApi.page({ model: STREAM_ID_A, first: 100 })).rejects.toThrow(
+      IndexQueryNotAvailableError
+    )
   })
 })