+++ packages/core/src/indexing/postgres/postgres-index-api.ts
@@ -1,14 +1,20 @@
 import { StreamID } from 'streamid/lib/stream-id.js'
 import type { BaseQuery, Pagination, Page } from '@ceramicnetwork/common'
 import type { DatabaseIndexApi, IndexStreamArgs } from '../database-index-api.js'
-import { initTables } from '../postgres/init-tables.js'
-import { InsertionOrder } from '../postgres/insertion-order.js'
+import { initTables, verifyTables } from './init-tables.js'
+import { InsertionOrder } from './insertion-order.js'
 import { asTableName } from '../as-table-name.util.js'
 import { Knex } from 'knex'
+import { IndexQueryNotAvailableError } from '../index-query-not-available.error.js'
+import { validTableStructure } from './migrations/mid-schema-verification.js'
 
 export class PostgresIndexApi implements DatabaseIndexApi {
-  private readonly insertionOrder: InsertionOrder
-  constructor(private readonly dbConnection: Knex, readonly modelsToIndex: Array<StreamID>) {
+  readonly insertionOrder: InsertionOrder
+  constructor(
+    private readonly dbConnection: Knex,
+    readonly modelsToIndex: Array<StreamID>,
+    private readonly allowQueriesBeforeHistoricalSync: boolean
+  ) {
     this.insertionOrder = new InsertionOrder(dbConnection)
   }
 
@@ -42,11 +48,20 @@ export class PostgresIndexApi implements DatabaseIndexApi {
   }
 
   async page(query: BaseQuery & Pagination): Promise<Page<StreamID>> {
+    // TODO(NET-1630) Throw if historical indexing is in progress
+    if (!this.allowQueriesBeforeHistoricalSync) {
+      throw new IndexQueryNotAvailableError(query.model)
+    }
     return this.insertionOrder.page(query)
+  }
+
+  async verify(validTableStructure: Object): Promise<void> {
+    await verifyTables(this.dbConnection, this.modelsToIndex, validTableStructure)
   }
 
   async init(): Promise<void> {
     await initTables(this.dbConnection, this.modelsToIndex)
+    await this.verify(validTableStructure)
   }
 
   async close(): Promise<void> {