+++ packages/core/src/indexing/sqlite/sqlite-index-api.ts
@@ -2,9 +2,11 @@ import type { StreamID } from '@ceramicnetwork/streamid'
 import type { BaseQuery, Pagination, Page } from '@ceramicnetwork/common'
 import type { Knex } from 'knex'
 import type { DatabaseIndexApi, IndexStreamArgs } from '../database-index-api.js'
-import { initTables } from './init-tables.js'
+import { initTables, verifyTables } from './init-tables.js'
 import { asTableName } from '../as-table-name.util.js'
 import { InsertionOrder } from './insertion-order.js'
+import { IndexQueryNotAvailableError } from '../index-query-not-available.error.js'
+import { validTableStructure } from './migrations/mid-schema-verfication.js'
 
 /**
  * Convert `Date` to SQLite `INTEGER`.
@@ -18,8 +20,13 @@ export function asTimestamp(input: Date | null | undefined): number | null {
 }
 
 export class SqliteIndexApi implements DatabaseIndexApi {
-  private readonly insertionOrder: InsertionOrder
-  constructor(private readonly dbConnection: Knex, readonly modelsToIndex: Array<StreamID>) {
+  readonly insertionOrder: InsertionOrder
+
+  constructor(
+    private readonly dbConnection: Knex,
+    readonly modelsToIndex: Array<StreamID>,
+    private readonly allowQueriesBeforeHistoricalSync: boolean
+  ) {
     this.insertionOrder = new InsertionOrder(dbConnection)
   }
 
@@ -53,11 +60,20 @@ export class SqliteIndexApi implements DatabaseIndexApi {
   }
 
   async page(query: BaseQuery & Pagination): Promise<Page<StreamID>> {
+    // TODO(NET-1630) Throw if historical indexing is in progress
+    if (!this.allowQueriesBeforeHistoricalSync) {
+      throw new IndexQueryNotAvailableError(query.model)
+    }
     return this.insertionOrder.page(query)
+  }
+
+  async verify(validTableStructure: Object): Promise<void> {
+    await verifyTables(this.dbConnection, this.modelsToIndex, validTableStructure)
   }
 
   async init(): Promise<void> {
     await initTables(this.dbConnection, this.modelsToIndex)
+    await this.verify(validTableStructure)
   }
 
   async close(): Promise<void> {