+++ packages/stream-model-instance-handler/src/model-instance-document-handler.ts
@@ -153,22 +153,15 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
       )
     }
 
-    const nextState = cloneDeep(state)
+    const oldContent = state.content
+    const newContent = jsonpatch.applyPatch(oldContent, payload.data).newDocument
+    await this._validateContent(context, metadata.model, newContent)
 
+    const nextState = cloneDeep(state)
     nextState.signature = SignatureStatus.SIGNED
     nextState.anchorStatus = AnchorStatus.NOT_REQUESTED
-
+    nextState.content = newContent
     nextState.log.push({ cid: commitData.cid, type: CommitType.SIGNED })
-
-    const oldContent = state.next?.content ?? state.content
-    const newContent = jsonpatch.applyPatch(oldContent, payload.data).newDocument
-
-    nextState.next = {
-      content: newContent,
-      metadata, // No way to update metadata for ModelInstanceDocument streams
-    }
-
-    await this._validateContent(context, metadata.model, newContent)
 
     return nextState
   }
@@ -188,26 +181,18 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
     StreamUtils.assertCommitLinksToState(state, commitData.commit)
 
     const proof = commitData.proof
-    state.log.push({
+    const newState = {
+      ...state,
+      anchorStatus: AnchorStatus.ANCHORED,
+      anchorProof: proof,
+    }
+    newState.log.push({
       cid: commitData.cid,
       type: CommitType.ANCHOR,
       timestamp: proof.blockTimestamp,
     })
-    let content = state.content
 
-    if (state.next?.content) {
-      content = state.next.content
-    }
-
-    delete state.next
-    delete state.anchorScheduledFor
-
-    return {
-      ...state,
-      content,
-      anchorStatus: AnchorStatus.ANCHORED,
-      anchorProof: proof,
-    }
+    return newState
   }
 
   /**