+++ packages/stream-model/src/model.ts
@@ -121,12 +121,13 @@ export class Model extends Stream {
 
   // The hardcoded "model" StreamID that all Model streams have in their metadata. This provides
   // a "model" StreamID that can be indexed to query the set of all published Models.
+  // The StreamID uses the "UNLOADABLE" StreamType, and has string representation: "kh4q0ozorrgaq2mezktnrmdwleo1d"
   static readonly MODEL: StreamID = (function () {
     const data = encode('model-v1')
     const multihash = multihashes.encode(data, 'identity')
     const digest = create(code, multihash)
     const cid = CID.createV1(code, digest)
-    return new StreamID(Model.STREAM_TYPE_ID, cid)
+    return new StreamID('UNLOADABLE', cid)
   })()
 
   private _isReadOnly = false