+++ packages/3id-did-resolver/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.4-rc.2](/compare/@ceramicnetwork/3id-did-resolver@2.1.4-rc.1...@ceramicnetwork/3id-did-resolver@2.1.4-rc.2) (2022-07-26)
+
+**Note:** Version bump only for package @ceramicnetwork/3id-did-resolver
+
+
+
+
+
+## [2.1.4-rc.1](/compare/@ceramicnetwork/3id-did-resolver@2.1.4-rc.0...@ceramicnetwork/3id-did-resolver@2.1.4-rc.1) (2022-07-14)
+
+**Note:** Version bump only for package @ceramicnetwork/3id-did-resolver
+
+
+
+
+
 ## [2.1.4-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/3id-did-resolver@2.1.3...@ceramicnetwork/3id-did-resolver@2.1.4-rc.0) (2022-07-06)
 
 **Note:** Version bump only for package @ceramicnetwork/3id-did-resolver