+++ packages/pinning-crust-backend/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.9-rc.2](/compare/@ceramicnetwork/pinning-crust-backend@2.0.9-rc.1...@ceramicnetwork/pinning-crust-backend@2.0.9-rc.2) (2022-07-26)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-crust-backend
+
+
+
+
+
+## [2.0.9-rc.1](/compare/@ceramicnetwork/pinning-crust-backend@2.0.9-rc.0...@ceramicnetwork/pinning-crust-backend@2.0.9-rc.1) (2022-07-14)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-crust-backend
+
+
+
+
+
 ## [2.0.9-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-crust-backend@2.0.8...@ceramicnetwork/pinning-crust-backend@2.0.9-rc.0) (2022-07-06)
 
 **Note:** Version bump only for package @ceramicnetwork/pinning-crust-backend