+++ packages/stream-caip10-link-handler/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1-rc.2](/compare/@ceramicnetwork/stream-caip10-link-handler@2.1.1-rc.1...@ceramicnetwork/stream-caip10-link-handler@2.1.1-rc.2) (2022-07-26)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-caip10-link-handler
+
+
+
+
+
+## [2.1.1-rc.1](/compare/@ceramicnetwork/stream-caip10-link-handler@2.1.1-rc.0...@ceramicnetwork/stream-caip10-link-handler@2.1.1-rc.1) (2022-07-14)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-caip10-link-handler
+
+
+
+
+
 ## [2.1.1-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-caip10-link-handler@2.1.0...@ceramicnetwork/stream-caip10-link-handler@2.1.1-rc.0) (2022-07-06)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-caip10-link-handler