+++ packages/stream-model/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0-rc.0](/compare/@ceramicnetwork/stream-model@0.3.1-rc.1...@ceramicnetwork/stream-model@0.4.0-rc.0) (2022-07-26)
+
+
+### Features
+
+* **stream-model:** Change Model model for indexing to use UNLOADABLE streamtype (#2326) cac8010, closes #2326
+
+
+
+
+
+## [0.3.1-rc.1](/compare/@ceramicnetwork/stream-model@0.3.1-rc.0...@ceramicnetwork/stream-model@0.3.1-rc.1) (2022-07-14)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-model
+
+
+
+
+
 ## [0.3.1-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-model@0.3.0...@ceramicnetwork/stream-model@0.3.1-rc.0) (2022-07-06)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-model