+++ packages/stream-tile/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0-rc.1](/compare/@ceramicnetwork/stream-tile@2.4.0-rc.0...@ceramicnetwork/stream-tile@2.4.0-rc.1) (2022-07-26)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile
+
+
+
+
+
+# [2.4.0-rc.0](/compare/@ceramicnetwork/stream-tile@2.3.1-rc.0...@ceramicnetwork/stream-tile@2.4.0-rc.0) (2022-07-14)
+
+
+### Features
+
+* **common:** Remove 'controllers' accessor from Stream base class (#2305) 05bf615, closes #2305
+
+
+
+
+
 ## [2.3.1-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile@2.3.0...@ceramicnetwork/stream-tile@2.3.1-rc.0) (2022-07-06)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-tile